--- ks43team03/src/main/resources/mapper/OfferMapper.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="ks43team03.mapper.OfferMapper">
-	
-	
-	<select id="getOfferList" resultType="map">
-		SELECT
-			a.area_nm,
-			f.facility_nm,
-			f.facility_address,
-			l.lesson_cd,
-			l.facility_goods_cd,
-			of.lesson_offer_cd,
-			of.offer_title,
-			of.offer_end_date,
-			of.offer_state
-		FROM
-			area AS  a
-			INNER join
-			facility AS f
-			on
-			a.area_cd = f.area_cd
-			INNER JOIN 
-			lesson AS l
-			on
-			f.facility_cd = l.facility_cd
-			INNER join
-			lesson_offer AS of
-			on
-			of.lesson_cd = l.lesson_cd
-	</select>
-</mapper>
-
