+++ ks43team03/src/main/resources/templates/offer/LessonOfferList.html
@@ -87,7 +87,7 @@
 									</ul>
 								</div>
 								<div class="product__item__text">
-									<h6><a href="#" th:text="${offer['offer_title']}"></a></h6>
+									<h6><a th:href="@{/offer/offerDetail(lessonOfferCd=${offer['lesson_offer_cd']})}" th:text="${offer['offer_title']}"></a></h6>
 									<div class="rating">
 										<i class="fa fa-star"></i>
 										<i class="fa fa-star"></i>
